feat(MON-142): fetch sensor metadata from fulcrum (#181)

+++ core/src/main/java/ch/swisstopo/monteis/core/modules/sensor/web/dto/nested/CoordinatesDto.java
@@ -2,4 +2,4 @@ package ch.swisstopo.monteis.core.modules.sensor.web.dto.nested;
 
 import jakarta.validation.constraints.NotNull;
 
-public record CoordinatesDto(@NotNull Integer x, @NotNull Integer y, @NotNull Integer z) {}
+public record CoordinatesDto(@NotNull Double x, @NotNull Double y, @NotNull Double z) {}
